email sends on confirmation
